BUILD: move mpscq to its own BUILD target so it can be used without gpr
