+++ config.w32
@@ -488,6 +488,7 @@ if (PHP_GRPC != "no") {
     "src\\core\\lib\\gprpp\\global_config_env.cc " +
     "src\\core\\lib\\gprpp\\host_port.cc " +
     "src\\core\\lib\\gprpp\\mpscq.cc " +
+    "src\\core\\lib\\gprpp\\mpscq_locked.cc " +
     "src\\core\\lib\\gprpp\\stat_posix.cc " +
     "src\\core\\lib\\gprpp\\stat_windows.cc " +
     "src\\core\\lib\\gprpp\\status_helper.cc " +
