+++ gRPC-C++.podspec
@@ -726,6 +726,7 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/match.h',
                       'src/core/lib/gprpp/memory.h',
                       'src/core/lib/gprpp/mpscq.h',
+                      'src/core/lib/gprpp/mpscq_locked.h',
                       'src/core/lib/gprpp/no_destruct.h',
                       'src/core/lib/gprpp/orphanable.h',
                       'src/core/lib/gprpp/overload.h',
@@ -1581,6 +1582,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/match.h',
                               'src/core/lib/gprpp/memory.h',
                               'src/core/lib/gprpp/mpscq.h',
+                              'src/core/lib/gprpp/mpscq_locked.h',
                               'src/core/lib/gprpp/no_destruct.h',
                               'src/core/lib/gprpp/orphanable.h',
                               'src/core/lib/gprpp/overload.h',
