+++ gRPC-Core.podspec
@@ -1146,6 +1146,8 @@ Pod::Spec.new do |s|
                       'src/core/lib/gprpp/memory.h',
                       'src/core/lib/gprpp/mpscq.cc',
                       'src/core/lib/gprpp/mpscq.h',
+                      'src/core/lib/gprpp/mpscq_locked.cc',
+                      'src/core/lib/gprpp/mpscq_locked.h',
                       'src/core/lib/gprpp/no_destruct.h',
                       'src/core/lib/gprpp/orphanable.h',
                       'src/core/lib/gprpp/overload.h',
@@ -2203,6 +2205,7 @@ Pod::Spec.new do |s|
                               'src/core/lib/gprpp/match.h',
                               'src/core/lib/gprpp/memory.h',
                               'src/core/lib/gprpp/mpscq.h',
+                              'src/core/lib/gprpp/mpscq_locked.h',
                               'src/core/lib/gprpp/no_destruct.h',
                               'src/core/lib/gprpp/orphanable.h',
                               'src/core/lib/gprpp/overload.h',
