+++ grpc.gemspec
@@ -1059,6 +1059,8 @@ Gem::Specification.new do |s|
   s.files += %w( src/core/lib/gprpp/memory.h )
   s.files += %w( src/core/lib/gprpp/mpscq.cc )
   s.files += %w( src/core/lib/gprpp/mpscq.h )
+  s.files += %w( src/core/lib/gprpp/mpscq_locked.cc )
+  s.files += %w( src/core/lib/gprpp/mpscq_locked.h )
   s.files += %w( src/core/lib/gprpp/no_destruct.h )
   s.files += %w( src/core/lib/gprpp/orphanable.h )
   s.files += %w( src/core/lib/gprpp/overload.h )
