+++ grpc.gyp
@@ -344,6 +344,7 @@
         'src/core/lib/gprpp/global_config_env.cc',
         'src/core/lib/gprpp/host_port.cc',
         'src/core/lib/gprpp/mpscq.cc',
+        'src/core/lib/gprpp/mpscq_locked.cc',
         'src/core/lib/gprpp/stat_posix.cc',
         'src/core/lib/gprpp/stat_windows.cc',
         'src/core/lib/gprpp/thd_posix.cc',
