+++ tools/doxygen/Doxyfile.core.internal
@@ -1834,6 +1834,8 @@ src/core/lib/gprpp/match.h \
 src/core/lib/gprpp/memory.h \
 src/core/lib/gprpp/mpscq.cc \
 src/core/lib/gprpp/mpscq.h \
+src/core/lib/gprpp/mpscq_locked.cc \
+src/core/lib/gprpp/mpscq_locked.h \
 src/core/lib/gprpp/no_destruct.h \
 src/core/lib/gprpp/orphanable.h \
 src/core/lib/gprpp/overload.h \
